requirements.txt
Pillow
tk
zipfile36